requirements.txt
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
astroid==2.15.8
asttokens==2.4.0
attrs==23.2.0
backcall==0.2.0
blinker==1.6.2
branca==0.6.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
dill==0.3.7
dnspython==2.4.2
email-validator==2.1.0.post1
executing==2.0.0
fastapi==0.104.1
Flask==2.3.3
folium==0.14.0
fonttools==4.43.1
gitdb==4.0.10
GitPython==3.1.34
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.4
iniconfig==2.0.0
ipykernel==6.25.2
ipython==8.16.1
isort==5.12.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.4.0
jupyter_core==5.4.0
kiwisolver==1.4.5
lazy-object-proxy==1.9.0
mando==0.7.1
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mccabe==0.7.0
nest-asyncio==1.5.8
numpy==1.26.1
orjson==3.9.10
packaging==23.1
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.10.0
pluggy==1.4.0
prompt-toolkit==3.0.39
psutil==5.9.6
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
Pygments==2.16.1
pylint==2.17.6
pyparsing==3.1.1
pyTelegramBotAPI==4.13.0
pytest==7.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-gitlab==3.15.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
radon==6.0.1
redis==5.0.1
requests==2.31.0
requests-toolbelt==1.0.0
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.23
stack-data==0.6.3
starlette==0.27.0
telebot==0.0.5
threadpoolctl==3.2.0
tomli==2.0.1
tomlkit==0.12.1
tornado==6.3.3
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.8
websockets==12.0
Werkzeug==2.3.7
wrapt==1.15.0
